# api/public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Benevolus",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/campaigns/{campaign}/withdrawals": {
            "get": {
                "tags": [
                    "Withdrawals"
                ],
                "summary": "List campaign withdrawals",
                "operationId": "listWithdrawals",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Withdrawals retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/WithdrawalResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Withdrawals"
                ],
                "summary": "Create withdrawal",
                "operationId": "createWithdrawal",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreWithdrawalRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Withdrawal request created successfully"
                    }
                }
            }
        },
        "/api/withdrawals/{withdrawal}": {
            "get": {
                "tags": [
                    "Withdrawals"
                ],
                "summary": "Get withdrawal by ID",
                "operationId": "getWithdrawal",
                "parameters": [
                    {
                        "name": "withdrawal",
                        "in": "path",
                        "description": "Withdrawal ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Withdrawal retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WithdrawalResource"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreWithdrawalRequest": {
                "required": [
                    "amountCents",
                    "pixKey",
                    "pixKeyType"
                ],
                "properties": {
                    "amountCents": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "pixKey": {
                        "type": "string"
                    },
                    "pixKeyType": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "CampaignResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "goal": {
                        "type": "number"
                    },
                    "raised": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "WithdrawalResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "campaignId": {
                        "type": "integer"
                    },
                    "amountCents": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    },
                    "pixKey": {
                        "type": "string"
                    },
                    "pixKeyType": {
                        "type": "string"
                    },
                    "paidAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "campaign": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/CampaignResource"
                            }
                        ],
                        "nullable": true,
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}